+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ This a informal changelog so i can keep track of what im doing.
 ## 2026-04-21
 
 - Added a button in settings to mass delete orphaned cbz files. (because tracking downloaded files properly is too annoying)
+- Add better provider listing for each series. You can now see the matches rebarr makes.
 
 ## 2026-04-19 - 2026-04-20